Add Amphion Malone decoder backend (used on i.MX8 QuadMax and QuadXPlus)
NOTE: Due to Malone driver requirements, the PTS is now required.
This is a breaking change, but unfortunately cannot be avoided.
Existing setups that run on Hantro-enabled hardware are not affected;
this is purely about cases when libimxvpuapi is run on QuadMax/QuadXPlus
i.MX8 SoCs.

This is the first working version of this backend.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>