amphion: Skip decoded frames with interpolated timestamps
This avoids the "assert(frame_context_index >= 0)" assertion that sometimes
happens when the timestamp manager inside the NXP Amphion Malone B0 driver
decides to interpolate entirely new timestamps. It is a workaround for this
V4L2 spec violation in that driver. In the absence of an option to turn
off the timestamp manager, the only other "clean" approach is to always
turn off frame reordering in the driver and do the reordering manually,
but this is a far more complex task. For now, skip frames with interpolated
timestamps. These typically do not occur often.

Also, since such interpolated frames unfortunately are a normal part of
decoding, change log levels from WARNING/ERROR to DEBUG.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>