imx8m-hantro: Reset decoder write_offset to fill_level when moving read_offset.
With the right timing, it's possible to have the read side all caught up and the
write side at the end of the buffer. This will cause the memmove() to move 0 bytes
and the write offset to be reduced by 0, leading to a SIGSEGV in the memcpy() as we
try to write after the buffer.

Signed-off-by: Doug Nazar <nazard@nazar.ca>