Implement first working version of VC8000E encoder integration
This adds support for the Hanto VC8000E encoder that can be found on the
i.MX8m plus. The code requires some restructuring for this - most notably,
the dummy encoder bits are factored out into a separate file, to be used
with the i.MX8m quad (which has no encoder).

The VC8000E encoder integration is not finished. h.264 encoding works
fine. h.265 encoding is not yet done. And, changing framerate mid-stream
is currently not supported.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>