vpu: encoder: Add imx_vpu_api_enc_get_encoded_frame_ext()
This is a preparation for reworked intra refresh support. In intra
refresh encoding, only the first frame is an I/IDR frame; all followup
frames are P frames. The start of an intra refresh interval qualifies
as a sync point. But since only P frames are present, callers callers
can't rely on the frame type to detect sync points. To fix this, the
imx_vpu_api_enc_get_encoded_frame_ext() is added to the API to be able
to communicate to callers that a sync point was produced.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>